[Bronze V] Title: 팩토리얼 2, Time: 124 ms, Memory: 14204 KB -BaekjoonHub

+++ 백준/Bronze/27433. 팩토리얼 2/팩토리얼 2.java
@@ -6,7 +6,7 @@ public class Main {
 	public static void main(String[] args) throws IOException{
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		
-		long n = Integer.parseInt(br.readLine());
+		int n = Integer.parseInt(br.readLine());
 		
 		long temp = 1;
